06.11.22

+++ src/J04_DongulerVeKararMekanizmalari/Pratik07_HarmonikSayilar.java
@@ -8,20 +8,13 @@ public class Pratik07_HarmonikSayilar {
  Java ile girilen sayının harmonik serisini bulan program yazacağız.
 
 Harmonik Seri Formülü :
-
-
-
  */
-
-
         Scanner input = new Scanner(System.in);
         int sayi;
         double harm = 0, unit = 0;
-
         System.out.println("Sayınızı giriniz :");
         sayi = input.nextInt();
-
-        for (int i = 1; i <=sayi; i++ ) {
+        for (double i = 1; i <=sayi; i++ ) {
             unit = (1.0 / i);
             harm += unit;
         }
